decouple toml details from chainspec type
