+++ node/src/data_migration.rs
@@ -10,8 +10,10 @@ use casper_types::{crypto, ProtocolVersion, PublicKey, SecretKey, Signature};
 
 use crate::{
     reactor::main_reactor::Config,
-    types::{chainspec, Chainspec, ChainspecRawBytes},
-    utils::{LoadError, Loadable, WithDir},
+    types::{Chainspec, ChainspecRawBytes},
+    utils::{
+        chain_specification::error::Error as LoadChainspecError, LoadError, Loadable, WithDir,
+    },
 };
 
 /// The name of the file for recording the new global state hash after a data migration.
@@ -80,7 +82,7 @@ pub(crate) enum Error {
 
     /// Error loading the chainspec.
     #[error("error loading chainspec: {0}")]
-    LoadChainspec(chainspec::Error),
+    LoadChainspec(LoadChainspecError),
 }
 
 #[derive(Serialize, Deserialize)]
